adk_agent/requirements.txt
google-adk
google-generativeai
absl-py
